renaming global navigation bar functionalities
