+++ src/test/java/com/wikia/webdriver/testcases/globalnavigationbar/Navigating.java
@@ -1,18 +1,18 @@
-package com.wikia.webdriver.testcases.globalnavigationtests;
+package com.wikia.webdriver.testcases.globalnavigationbar;
 
 import com.wikia.webdriver.common.core.Assertion;
 import com.wikia.webdriver.common.logging.PageObjectLogging;
 import com.wikia.webdriver.common.templates.NewTestTemplate;
 import com.wikia.webdriver.pageobjectsfactory.pageobject.HomePageObject;
-
 import com.wikia.webdriver.pageobjectsfactory.pageobject.globalnav.GlobalNavigationPageObject;
+
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 @Test(groups = {"WikiaLogo", "GlobalNav"})
-public class GlobalNavigationWikiaLogo extends NewTestTemplate {
+public class Navigating extends NewTestTemplate {
 
   private final String EN_COMMUNITY = "muppet";
   private final String FANDOM_URL = "fandom.wikia";
@@ -49,5 +49,15 @@ public class GlobalNavigationWikiaLogo extends NewTestTemplate {
     PageObjectLogging.log("CHECK URL", "Expected: " + urlBuilder.getUrlForWiki(FANDOM_URL),
             new WebDriverWait(driver, 10).until(ExpectedConditions.urlContains(urlBuilder
                     .getUrlForWiki(FANDOM_URL))));
+  }
+
+  @Test(groups = {"TestHubLinksInGlobalNav_001"},
+      enabled = false)
+  public void TestHubLinksInGlobalNav_001_clickHubsLinks() {
+    GlobalNavigationPageObject globalNav = new HomePageObject(driver).getGlobalNavigation();
+
+    for (GlobalNavigationPageObject.Hub hubName : GlobalNavigationPageObject.Hub.values()) {
+      globalNav.openHub(hubName);
+    }
   }
 }
