--- src/test/java/com/wikia/webdriver/testcases/globalnavigationtests/GlobalNavigationHubLinks.java
@@ -1,21 +1,0 @@
-package com.wikia.webdriver.testcases.globalnavigationtests;
-
-import com.wikia.webdriver.common.templates.NewTestTemplate;
-import com.wikia.webdriver.pageobjectsfactory.pageobject.HomePageObject;
-import com.wikia.webdriver.pageobjectsfactory.pageobject.globalnav.GlobalNavigationPageObject;
-
-import org.testng.annotations.Test;
-
-@Test(groups = {"HubLinksInGlobalNav", "GlobalNav"})
-public class GlobalNavigationHubLinks extends NewTestTemplate {
-
-  @Test(groups = {"TestHubLinksInGlobalNav_001"},
-  enabled = false)
- public void TestHubLinksInGlobalNav_001_clickHubsLinks() {
-    GlobalNavigationPageObject globalNav = new HomePageObject(driver).getGlobalNavigation();
-
-    for (GlobalNavigationPageObject.Hub hubName : GlobalNavigationPageObject.Hub.values()) {
-      globalNav.openHub(hubName);
-    }
-  }
-}
