+++ src/test/java/com/wikia/webdriver/testsuites/testSuite.xml
@@ -92,13 +92,9 @@
             <class name="com.wikia.webdriver.testcases.forumtests.ForumEditModeTests"/>
             <class name="com.wikia.webdriver.testcases.forumtests.ForumThreadTests"/>
             <class name="com.wikia.webdriver.testcases.googleconnect.GoogleConnect"/>
-            <class name="com.wikia.webdriver.testcases.globalnavigationtests.GameStarLogo"/>
-            <class name="com.wikia.webdriver.testcases.globalnavigationtests.GlobalNavigationHubLinks"/>
-            <class name="com.wikia.webdriver.testcases.globalnavigationtests.GlobalNavigationScrollAction"/>
-            <class name="com.wikia.webdriver.testcases.globalnavigationtests.GlobalNavigationSearch"/>
-            <class name="com.wikia.webdriver.testcases.globalnavigationtests.GlobalNavigationWikiaLogo"/>
-            <class name="com.wikia.webdriver.testcases.globalnavigationtests.ExploreWikiaDropdown"/>
-            <class name="com.wikia.webdriver.testcases.globalnavigationtests.HubsLinks"/>
+            <class name="com.wikia.webdriver.testcases.globalnavigationbar.Loading"/>
+            <class name="com.wikia.webdriver.testcases.globalnavigationbar.Searching"/>
+            <class name="com.wikia.webdriver.testcases.globalnavigationbar.Navigating"/>
             <class name="com.wikia.webdriver.testcases.heromoduletests.HeroModuleTests"/>
             <class name="com.wikia.webdriver.testcases.hubstests.HubsTests"/>
             <class name="com.wikia.webdriver.testcases.imageservingtests.ImageServingTests"/>
